refactor: startScreen 디렉토리 생성
- min27604/baseball/#10
