+++ FE/react-baseball/src/components/startScreen/Game.jsx
