+++ FE/react-baseball/src/components/startScreen/GameList.jsx
