+++ FE/react-baseball/src/components/startScreen/Team.jsx
