+++ FE/react-baseball/src/components/startScreen/Title.jsx
